Emailing - on
+EmailNotifier(updated Supply)
+MailContriller
+UserLocale(updated User)

+++ src/main/java/com/stardusted1/Sendicate/app/core/repositories/DeliverymanRepository.java
@@ -3,5 +3,5 @@ package com.stardusted1.Sendicate.app.core.repositories;
 import com.stardusted1.Sendicate.app.core.users.Deliveryman;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeliverymanRepository extends CrudRepository<Deliveryman, Integer> {
+public interface DeliverymanRepository extends CrudRepository<Deliveryman, Long> {
 }
