+++ src/main/java/com/stardusted1/Sendicate/app/core/repositories/ProviderRepository.java
@@ -3,5 +3,5 @@ package com.stardusted1.Sendicate.app.core.repositories;
 import com.stardusted1.Sendicate.app.core.users.Provider;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProviderRepository extends CrudRepository<Provider,Integer> {
+public interface ProviderRepository extends CrudRepository<Provider,Long> {
 }
